Refactoring to write star files, now picking test running (both 3.0/3.1)
