+++ relion/protocols/protocol_autopick_log.py
@@ -30,17 +30,12 @@ from os.path import relpath
 import pyworkflow.protocol.params as params
 from pyworkflow.protocol import STEPS_SERIAL
 from pyworkflow.em.protocol import ProtParticlePickingAuto
-from pyworkflow.em.constants import ALIGN_NONE
-import pyworkflow.utils as pwutils
-import pyworkflow.em.metadata as md
 
 import relion
-from .protocol_base import ProtRelionBase
-from relion.convert import (writeSetOfMicrographs, readSetOfCoordinates,
-                            micrographToRow)
+from .protocol_autopick import ProtRelionAutopickBase
 
 
-class ProtRelionAutopickLoG(ProtParticlePickingAuto, ProtRelionBase):
+class ProtRelionAutopickLoG(ProtRelionAutopickBase):
     """
     This Relion protocol uses 'relion_autopick' program for the
     Laplacian of Gaussian (LoG) option.
@@ -95,7 +90,7 @@ class ProtRelionAutopickLoG(ProtParticlePickingAuto, ProtRelionBase):
                             'size. Give a negative value to skip downscaling.')
 
         group.addParam('threshold', params.FloatParam, default=0,
-                       label='Adjust default threshold',
+                       label='Adjust default threshold (stddev)',
                        help='Use this to pick more (negative number -> lower '
                             'threshold) or less (positive number -> higher '
                             'threshold) particles compared to the default '
@@ -137,58 +132,20 @@ class ProtRelionAutopickLoG(ProtParticlePickingAuto, ProtRelionBase):
             self.threshold.get()
         ]
 
-    def _pickMicrographsFromStar(self, micStarFile, params,
+    def _pickMicrographsFromStar(self, micStarFile, cwd, params,
                                  minDiameter, maxDiameter, threshold):
         """ Launch the 'relion_autopick' for micrographs in the inputStarFile.
          If the input set of complete, the star file will contain all the
          micrographs. If working in streaming, it will be only one micrograph.
         """
-        params += ' --i %s' % relpath(micStarFile, self.getWorkingDir())
+        params += ' --i %s' % relpath(micStarFile, cwd)
         params += ' --LoG_diam_min %0.3f' % minDiameter
         params += ' --LoG_diam_max %0.3f' % maxDiameter
         params += ' --LoG_adjust_threshold %0.3f' % threshold
 
         program = self._getProgram('relion_autopick')
 
-        self.runJob(program, params, cwd=self.getWorkingDir())
-
-    def _pickMicrograph(self, mic, *args):
-        """ This method should be invoked only when working in streaming mode.
-        """
-        micRow = md.Row()
-        self._preprocessMicrographRow(mic, micRow)
-        micrographToRow(mic, micRow)
-        self._postprocessMicrographRow(mic, micRow)
-        self._pickMicrographsFromStar(self._getMicStarFile(mic), *args)
-
-    def _pickMicrographList(self, micList, *args):
-        micStar = self._getPath('input_micrographs_%s-%s.star' %
-                                (micList[0].strId(), micList[-1].strId()))
-        writeSetOfMicrographs(micList, micStar,
-                              alignType=ALIGN_NONE,
-                              preprocessImageRow=self._preprocessMicrographRow)
-        self._pickMicrographsFromStar(micStar, *args)
-
-    def _createSetOfCoordinates(self, micSet, suffix=''):
-        """ Override this method to set the box size. """
-        coordSet = ProtParticlePickingAuto._createSetOfCoordinates(self, micSet,
-                                                                   suffix=suffix)
-        coordSet.setBoxSize(self.getBoxSize())
-
-        return coordSet
-
-    def readCoordsFromMics(self, workingDir, micList, coordSet):
-        """ Parse back the output star files and populate the SetOfCoordinates.
-        """
-        template = self._getExtraPath("%s_autopick.star")
-        starFiles = [template % pwutils.removeBaseExt(mic.getFileName())
-                     for mic in micList]
-        readSetOfCoordinates(coordSet, starFiles, micList)
-
-    # -------------------------- STEPS functions -------------------------------
-    def autopickStep(self, micStarFile, params, *args):
-        """ This method is used from the wizard to optimize the parameters. """
-        self._pickMicrographsFromStar(micStarFile, *args)
+        self.runJob(program, params, cwd=cwd)
 
     # -------------------------- INFO functions --------------------------------
     def _validate(self):
@@ -210,68 +167,3 @@ class ProtRelionAutopickLoG(ProtParticlePickingAuto, ProtRelionBase):
     def getBoxSize(self):
         """ Return a reasonable box-size in pixels. """
         return self.boxSize.get()
-                
-    def getInputMicrographsPointer(self):
-        return self.inputMicrographs
-
-    def getInputMicrographs(self):
-        return self.getInputMicrographsPointer().get()
-
-    def getMicrographList(self):
-        """ Return the list of micrographs (either a subset or the full set)
-        that will be used for optimizing the parameters or the picking.
-        """
-        # Use all micrographs only when going for the full picking
-        return self.getInputMicrographs()
-
-    def getCoordsDir(self):
-        return self._getTmpPath('xmipp_coordinates')
-
-    def _writeXmippCoords(self, coordSet):
-        micSet = self.getInputMicrographs()
-        coordPath = self._getTmpPath('xmipp_coordinates')
-        pwutils.cleanPath(coordPath)
-        pwutils.makePath(coordPath)
-        import pyworkflow.em.packages.xmipp3 as xmipp3
-        micPath = micSet.getFileName()
-        xmipp3.writeSetOfCoordinates(coordPath, coordSet, ismanual=False)
-        return micPath, coordPath
-
-    def writeXmippOutputCoords(self):
-        return self._writeXmippCoords(self.outputCoordinates)
-
-    def writeXmippCoords(self):
-        """ Write the SetOfCoordinates as expected by Xmipp
-        to be displayed with its GUI.
-        """
-        micSet = self.getInputMicrographs()
-        coordSet = self._createSetOfCoordinates(micSet)
-        coordSet.setBoxSize(self.getBoxSize())
-        starFiles = [self._getExtraPath(pwutils.removeBaseExt(mic.getFileName())
-                                        + '_autopick.star') for mic in micSet]
-        readSetOfCoordinates(coordSet, starFiles)
-        return self._writeXmippCoords(coordSet)
-
-    def _preprocessMicrographRow(self, img, imgRow):
-        # Temporarly convert the few micrographs to tmp and make sure
-        # they are in 'mrc' format
-        # Get basename and replace extension by 'mrc'
-        newName = pwutils.replaceBaseExt(img.getFileName(), 'mrc')
-        newPath = self._getExtraPath(newName)
-
-        # If the micrographs are in 'mrc' format just create a link
-        # if not, convert to 'mrc'
-        if img.getFileName().endswith('mrc'):
-            pwutils.createLink(img.getFileName(), newPath)
-        else:
-            self._ih.convert(img, newPath)
-        # The command will be launched from the working dir
-        # so, let's make the micrograph path relative to that
-        img.setFileName(os.path.join('extra', newName))
-
-    def _postprocessMicrographRow(self, img, imgRow):
-        imgRow.writeToFile(self._getMicStarFile(img))
-
-    def _getMicStarFile(self, mic):
-        micBase = pwutils.replaceBaseExt(mic.getFileName(), 'star')
-        return self._getExtraPath(micBase)
